% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_trial_hte_data.R
\docType{data}
\name{example_trial_hte_data}
\alias{example_trial_hte_data}
\title{Simulated trial data (with HTE)}
\format{
An object of class \code{data.frame} with 6000 rows and 16 columns.
}
\source{
Rigdon, J., Baiocchi, M. & Basu, S. Preventing false discovery of
heterogeneous treatment effect subgroups in randomized trials. Trials 19,
382 (2018). https://doi.org/10.1186/s13063-018-2774-5
}
\usage{
example_trial_hte_data
}
\description{
Simulated trial data (with HTE)
}
\keyword{datasets}
