% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_averaged_matches.R
\name{get_averaged_matches}
\alias{get_averaged_matches}
\title{Get averaged matches}
\usage{
get_averaged_matches(
  matched_data,
  treatment,
  outcome,
  covariates,
  match_id = "match_id",
  weight = "wt"
)
}
\arguments{
\item{matched_data}{Filtered matched data set obtained from \code{\link[=get_filtered_matches]{get_filtered_matches()}}}

\item{treatment}{Character. Name of the treatment column in \code{data}. Currently
assumes \code{1} for treatment and \code{0} for control.}

\item{outcome}{Character. Name of the outcome column in \code{data}.}

\item{covariates}{Character vector. Vector of covariate column names
from \code{data} to use in match}

\item{match_id}{Character. The column name in \code{matched_data} that contains
the matching identifier. Default is "\code{match_id}".}

\item{weight}{Character. The column name in \code{matched_data} that contains
the weight. Default is "\code{wt}".}
}
\value{
A data frame of averaged matches
}
\description{
Get averaged matches
}
\examples{
matches <- get_filtered_matches(
  example_trial_data,
  treatment = "Z",
  covariates = c("age", "stage4", "site", "ecog", "prevTrt", "diseaseFree")
)
get_averaged_matches(
  matches,
  treatment = "Z",
  outcome = "Y",
  covariates = c("age", "stage4", "site", "ecog", "prevTrt", "diseaseFree"))
}
