% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_filtered_matches.R
\name{get_filtered_matches}
\alias{get_filtered_matches}
\title{Get filtered matches}
\usage{
get_filtered_matches(
  data,
  treatment,
  covariates,
  id = NULL,
  weight = "wt",
  ind_categorical = NULL
)
}
\arguments{
\item{data}{Data frame.}

\item{treatment}{Character. Name of the treatment column in \code{data}. Currently
assumes \code{1} for treatment and \code{0} for control.}

\item{covariates}{Character vector. Vector of covariate column names
from \code{data} to use in match}

\item{id}{Character. Name of the id column in \code{data}}

\item{weight}{Character. The name that should be given to the variable
containing the weights in the output data frame. Default is "\code{wt}".}

\item{ind_categorical}{Logical vector. (Optional) Indicator of which
.       \code{covariates} are categorical.}
}
\value{
A data frame of filtered matches
}
\description{
Get filtered matches
}
\examples{
get_filtered_matches(
  example_trial_data,
  treatment = "Z",
  covariates = c("age", "stage4", "site", "ecog", "prevTrt", "diseaseFree")
)
}
