% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mahalanobis_mat.R
\name{get_mahalanobis_mat}
\alias{get_mahalanobis_mat}
\title{Create rank based Mahalanobis distance matrix}
\usage{
get_mahalanobis_mat(X)
}
\arguments{
\item{X}{Matrix or dataframe to calculate distances with}
}
\value{
Rank based Mahalanobis distance matrix
}
\description{
Create rank based Mahalanobis distance matrix
}
\examples{
get_mahalanobis_mat(mtcars)
}
