% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mtree.R
\name{get_mtree}
\alias{get_mtree}
\title{Function to obtain decision tree}
\usage{
get_mtree(
  data,
  covariates,
  survivial = FALSE,
  l = -Inf,
  r = Inf,
  weight = NULL,
  delta = "delta",
  bins = 4
)
}
\arguments{
\item{data}{Data frame with averaged matches (obtained from \code{\link[=get_averaged_matches]{get_averaged_matches()}})}

\item{covariates}{Character vector. Vector of covariate column names
from \code{data} to use in match}

\item{survivial}{Logical. Indicator whether this is a survival analysis}

\item{l}{Numeric. Left}

\item{r}{Numeric. Right}

\item{weight}{Character. The name that should be given to the variable
containing the weights in the output data frame. Default is "\code{wt}".}

\item{delta}{Character. The name that should be given to the delta variable
in the output data frame. Default is "\code{delta}".}

\item{bins}{Numeric. The number of bins to use to categorize the weights.
Default: \code{4}.}
}
\value{
A decision tree object
}
\description{
Function to obtain decision tree
}
\examples{
covariates <- c("age", "stage4", "site", "ecog", "prevTrt", "diseaseFree")
matches <- get_filtered_matches(
  example_trial_data,
  treatment = "Z",
  covariates = covariates
)
m <- get_averaged_matches(
  matches,
  treatment = "Z",
  outcome = "Y",
  covariates = covariates
  )
get_mtree(m, covariates)
}
