% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mtree_stats.R
\name{get_mtree_stats}
\alias{get_mtree_stats}
\title{Calculate Effect Statistics}
\usage{
get_mtree_stats(data, treatment, outcome, covariates, tree, cens_var = NULL)
}
\arguments{
\item{data}{Data frame.}

\item{treatment}{Character. Name of the treatment column in \code{data}}

\item{outcome}{Character. Name of the outcome column in \code{data}}

\item{covariates}{Character vector. Vector of covariate column names
from \code{data} to use in match}

\item{tree}{Decision tree obtained from \code{\link[=get_mtree]{get_mtree()}}.}

\item{cens_var}{Character. Name of a censoring column in \code{data}. If none
exists, leave \code{NULL}. Default: \code{NULL}}
}
\value{
Data frame with effect estimates
}
\description{
Calculate Effect Statistics
}
\examples{
covariates <- c("age", "stage4", "site", "ecog", "prevTrt", "diseaseFree")
matches <- get_filtered_matches(
  example_trial_data,
  treatment = "Z",
  covariates = covariates
)
m <- get_averaged_matches(
  matches,
  treatment = "Z",
  outcome = "Y",
  covariates = covariates
  )
tree <- get_mtree(m, covariates)
get_mtree_stats(example_trial_data,
  treatment = "Z",
  outcome = "Y",
  covariates = covariates,
  tree = tree)
}
