% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtree_input}
\alias{mtree_input}
\title{Example mtree dataset.}
\format{A list of parameters to build a mutation tree.}
\usage{
data(mtree_input)
}
\description{
Example mtree dataset.
}
\examples{
data(mtree_input)
mtree_input

}
\keyword{datasets}
