% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_icon.R
\name{plot_icon}
\alias{plot_icon}
\title{Plot the information transfe among clones in icon format (tiny).}
\usage{
plot_icon(x, node_palette = colorRampPalette(RColorBrewer::brewer.pal(n =
  9, "Set1")), ...)
}
\arguments{
\item{x}{An \code{mtree} tree.}

\item{node_palette}{A function that can return, for an input number,
a number of colours.}

\item{...}{Other parameters, not used in this case.}
}
\value{
A \code{ggplot} object for the plot.
}
\description{
The information transfer of a tree is the set of orderings
associated to the internal annotated driver events.

This function plots the clones with drivers (so their orderings) following a topological
sort of the node of the corresponding clone tree, and in tiny
icon format. Some graohics changes with respect ot a standard plot.
}
\examples{
data(mtree_input)

x = mtrees(
mtree_input$binary_clusters, 
mtree_input$drivers,
mtree_input$samples,
mtree_input$patient,
mtree_input$sspace.cutoff,
mtree_input$n.sampling,
mtree_input$store.max
)

plot_icon(x[[1]])
}
