% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_information_transfer.R
\name{plot_information_transfer}
\alias{plot_information_transfer}
\title{Plot the information transfer of a tree (drivers ordering).}
\usage{
plot_information_transfer(x,
  node_palette = colorRampPalette(RColorBrewer::brewer.pal(n = 9,
  "Set1")), tree_layout = "tree", ...)
}
\arguments{
\item{x}{A \code{mtree} tree.}

\item{node_palette}{A function that can return, for an input number,
a number of colours.}

\item{tree_layout}{Layout of this model, as of \code{ggraph}.}

\item{...}{Other parameters, not used in this case.}
}
\value{
A \code{ggplot} object for the plot.
}
\description{
The information transfer of the tree is the set of orderings
associated to the internal annotated driver events.

This function plots these orderings following a topological
sort of the node of the corresponding clone tree.
}
\examples{
data(mtree_input)

x = mtrees(
mtree_input$binary_clusters, 
mtree_input$drivers,
mtree_input$samples,
mtree_input$patient,
mtree_input$sspace.cutoff,
mtree_input$n.sampling,
mtree_input$store.max
)

plot_information_transfer(x[[1]])

# Change layout -- use igraph's "kk" layout
plot_information_transfer(x[[1]], tree_layout = 'kk')
}
