% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Suppes.R
\name{plot_Suppes}
\alias{plot_Suppes}
\title{Plot Suppes' conditions for this tree}
\usage{
plot_Suppes(x, ...)
}
\arguments{
\item{x}{An \code{mtree} tree.}

\item{...}{Extra parameters, not used.}

\item{patient}{A patient id.}
}
\value{
A \code{ggplot} plot.
}
\description{
This function creates a joint plot assembled with
\code{cowplot} integrating two tile plots, one reporting which
conditions by Suppes' are satisfied, and the other reporting the
numbers used in the evaluation.
}
\examples{
data(mtree_input)

x = mtrees(
mtree_input$binary_clusters, 
mtree_input$drivers,
mtree_input$samples,
mtree_input$patient,
mtree_input$sspace.cutoff,
mtree_input$n.sampling,
mtree_input$store.max
)

plot_Suppes(x[[1]])
}
