% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_print.R
\name{print.mtree}
\alias{print.mtree}
\title{Print a \code{mtree} tree.}
\usage{
\method{print}{mtree}(x, ...)
}
\arguments{
\item{x}{A \code{mtree} tree.}

\item{...}{Extra S3 parameters}
}
\value{
Nothing
}
\description{
Print a \code{mtree} tree.
}
\details{
Print a summary for a \code{mtree} object, which includes a
shell-frinedly layout and other information.
}
\examples{
data(mtree_input)

x = mtrees(
mtree_input$binary_clusters, 
mtree_input$drivers,
mtree_input$samples,
mtree_input$patient,
mtree_input$sspace.cutoff,
mtree_input$n.sampling,
mtree_input$store.max
)

print(x[[1]])
}
