% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_summary.R
\name{summary.mtree}
\alias{summary.mtree}
\title{Summary of a \code{"mtree"} object.}
\usage{
\method{summary}{mtree}(x, ...)
}
\arguments{
\item{x}{An \code{mtree} tree.}

\item{...}{Extra parameters}
}
\value{
None.
}
\description{
Reports some summary statistics for a mutation tree.
}
\examples{
data(mtree_input)

x = mtrees(
mtree_input$binary_clusters, 
mtree_input$drivers,
mtree_input$samples,
mtree_input$patient,
mtree_input$sspace.cutoff,
mtree_input$n.sampling,
mtree_input$store.max
)

summary(x[[1]])
}
